Sequelize non-admin and token refresh
