+++ typescript/sequelize/package.json
@@ -3,13 +3,11 @@
     "version": "1.0.0",
     "description": "How To",
     "main": "index.js",
-    "type": "commonjs",
+    "type": "module",
     "scripts": {
         "build": "tsc",
-        "clean": "rm -rf build && rm -rf dist",
-        "start": "node dist/src/index.js",
-        "dev": "npm run clean && npm run build && npm run start",
-        "test": "npm run clean && npm run build && jest dist/test"
+        "start": "node dist/index.js",
+        "dev": "npm run build && npm run start"
     },
     "dependencies": {
         "@aws-sdk/dsql-signer": "^3.705.0",
@@ -19,8 +17,6 @@
         "sequelize": "^6.37.5"
     },
     "devDependencies": {
-        "@types/jest": "^29.5.14",
-        "jest": "^29.7.0",
         "typescript": "^5.0.0"
     }
 }
