--- typescript/sequelize/test/index.spec.ts
@@ -1,9 +1,0 @@
-import { runExamples } from "../src/sequelizeExample.js"
-
-const SECONDS = 1000;
-
-describe("Index module", () => {
-    test("Runs all the example code", async () => {
-        await runExamples();
-    }, 60 * SECONDS)
-})
