+++ typescript/sequelize/tsconfig.json
@@ -1,9 +1,13 @@
 {
   "compilerOptions": {
-    "target": "ES6",
-    "module": "node16",
-    "moduleResolution": "node16",
-    "outDir": "dist",
+    "target": "es2016",
+    "module": "ES2020",
+    "moduleResolution": "node",
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "outDir": "./dist",
+    "rootDir": "./src",
   },
-  "include": ["./src/**/*", "./test/**/*"]
+  "include" : ["src/**/*"]
 }
